I put the image in the answerAdapter so the answers are not hidden by the image
