+++ app/src/main/res/layout/fragment_my_questions.xml
@@ -14,8 +14,9 @@
         <TextView
             android:id="@+id/course_title_text_view"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:textSize="20sp"
+            android:text="My Questions"
             android:textStyle="bold"
             android:textColor="@android:color/white"
             android:background="@color/main_highlight"
@@ -49,8 +50,7 @@
             android:layout_height="wrap_content"
             tools:listitem="@layout/my_questions_item"
             android:clickable="true"
-            android:layout_marginLeft="10dp"
-            android:layout_marginRight="10dp"
+
             android:layout_marginBottom="10dp"/>
 
     </LinearLayout>
