+++ app/src/main/res/layout/question_details_header_item.xml
@@ -12,4 +12,12 @@
         android:textSize="30sp"
         android:id="@+id/qdetails_question_content"/>
 
+    <ImageView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/image_question"
+        android:layout_weight="1"
+        android:contentDescription="@string/image" />
+
+
 </RelativeLayout>
